delete, save, exit
