+++ dart_application_2/bin/college_app.dart
@@ -1,6 +1,7 @@
+import 'dart:io';
+
 import 'student.dart';
 import 'task.dart';
-import 'util.dart';
 
 class CollegeApp {
   static void startApp() {
@@ -12,10 +13,11 @@ class CollegeApp {
       print("chose an option");
 
       // displays the menu
-      Util.displayOption();
+      print("1: add user details\n2:display user details\n3:" +
+          "delete user details\n4:save user details\n5:exit");
 
       // chose one option from menu
-      option = Util.inputOption();
+      option = int.parse(stdin.readLineSync()!);
 
       // perform the task for given option and save the reult in list
       studentList = Task.performTask(option, studentList);
@@ -23,5 +25,8 @@ class CollegeApp {
       // sort the list by FullName and then Roll Number
       studentList.sort();
     } while (option != 5);
+    print("enter 4 to save, 5 to exit");
+    option = int.parse(stdin.readLineSync()!);
+    studentList = Task.performTask(option, studentList);
   }
 }
