+++ dart_application_2/bin/main.dart
@@ -1,5 +1,6 @@
 import 'college_app.dart';
 
 void main() {
+  // starting point of app
   CollegeApp.startApp();
 }
